fix(plugins): move plugin ownership handling to Extract in fs.go
